Cleaning up binary_map impl
